// src/app/models/material-price.model.ts
export interface MaterialPrice {
  id: string;
  material: string;
  id_material_type: number;
  coverage_sq: string;
  coverage_lf?: any;
  id_unit_buy: number;
  unit: string;
  abbreviation: string;
  id_trademark: number;
  trademark: string;
  id_supplier: number;
  material_type: string;
  bidsheet_order: number;
  id_material_category: number;
  category_bid_sheet_order: number;
  id_price_list: string;
  cost: string;
  id_material_price_list: string;
  start_date: string;
  supplier: string;
}